requirements.txt
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==2.0.12
idna==3.4
lxml==4.9.2
Pillow==8.4.0
python-docx==0.8.11
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.15
